+++ include/GafferVDB/Interrupter.h
@@ -47,8 +47,7 @@ class Interrupter {
 	public:
 
 		Interrupter(const IECore::Canceller *canceller)
-		:	m_canceller(canceller),
-			m_interrupted(false)
+			:	m_canceller( canceller )
 		{
 		}
 
@@ -62,27 +61,12 @@ class Interrupter {
 
 		bool wasInterrupted( int percent = -1 )
 		{
-			if ( m_interrupted )
-			{
-				return true;
-			}
-
-			// todo this a a problem installing a exception handler
-			// per a call to this function.
-			try
-			{
-				IECore::Canceller::check( m_canceller );
-			}
-			catch( const IECore::Cancelled& )
-			{
-				m_interrupted = true;
-			}
-
-			return m_interrupted;
+			return m_canceller && m_canceller->cancelled();
 		}
+
 	private:
+
 		const IECore::Canceller* m_canceller;
-		bool m_interrupted;
 
 };