+++ src/GafferVDB/LevelSetToMesh.cpp
@@ -175,7 +175,7 @@ GAFFER_NODE_DEFINE_TYPE( LevelSetToMesh );
 size_t LevelSetToMesh::g_firstPlugIndex = 0;
 
 LevelSetToMesh::LevelSetToMesh( const std::string &name )
-	:	SceneElementProcessor( name, IECore::PathMatcher::NoMatch )
+	:	Deformer( name )
 {
 	storeIndexOfNextChild( g_firstPlugIndex );
 
@@ -199,7 +199,6 @@ const Gaffer::StringPlug *LevelSetToMesh::gridPlug() const
 	return  getChild<StringPlug>( g_firstPlugIndex );
 }
 
-
 Gaffer::FloatPlug *LevelSetToMesh::isoValuePlug()
 {
 	return getChild<FloatPlug>( g_firstPlugIndex + 1);
@@ -220,42 +219,27 @@ const Gaffer::FloatPlug *LevelSetToMesh::adaptivityPlug() const
 	return getChild<FloatPlug>( g_firstPlugIndex + 2 );
 }
 
-void LevelSetToMesh::affects( const Gaffer::Plug *input, AffectedPlugsContainer &outputs ) const
+bool LevelSetToMesh::affectsProcessedObject( const Gaffer::Plug *input ) const
 {
-	SceneElementProcessor::affects( input, outputs );
-
-	if(
+	return
 		input == isoValuePlug() ||
 		input == adaptivityPlug() ||
 		input == gridPlug()
-	)
-	{
-		outputs.push_back( outPlug()->objectPlug() );
-	}
-
-	if( input == isoValuePlug() )
-	{
-		outputs.push_back( outPlug()->boundPlug() );
-	}
-}
-
-bool LevelSetToMesh::processesObject() const
-{
-	return true;
+	;
 }
 
 void LevelSetToMesh::hashProcessedObject( const ScenePath &path, const Gaffer::Context *context, IECore::MurmurHash &h ) const
 {
-	SceneElementProcessor::hashProcessedObject( path, context, h );
+	Deformer::hashProcessedObject( path, context, h );
 
 	gridPlug()->hash( h );
 	isoValuePlug()->hash( h );
 	adaptivityPlug()->hash( h );
 }
 
-IECore::ConstObjectPtr LevelSetToMesh::computeProcessedObject( const ScenePath &path, const Gaffer::Context *context, IECore::ConstObjectPtr inputObject ) const
+IECore::ConstObjectPtr LevelSetToMesh::computeProcessedObject( const ScenePath &path, const Gaffer::Context *context, const IECore::Object *inputObject ) const
 {
-	const VDBObject *vdbObject = runTimeCast<const VDBObject>( inputObject.get() );
+	const VDBObject *vdbObject = runTimeCast<const VDBObject>( inputObject );
 	if( !vdbObject )
 	{
 		return inputObject;
@@ -271,19 +255,23 @@ IECore::ConstObjectPtr LevelSetToMesh::computeProcessedObject( const ScenePath &
 	return volumeToMesh( grid, isoValuePlug()->getValue(), adaptivityPlug()->getValue() );
 }
 
-bool LevelSetToMesh::processesBound() const
+bool LevelSetToMesh::affectsProcessedObjectBound( const Gaffer::Plug *input ) const
 {
-	return true;
+	return input == inPlug()->boundPlug() || input == isoValuePlug();
 }
 
-void LevelSetToMesh::hashProcessedBound( const ScenePath &path, const Gaffer::Context *context, IECore::MurmurHash &h ) const
+void LevelSetToMesh::hashProcessedObjectBound( const ScenePath &path, const Gaffer::Context *context, IECore::MurmurHash &h ) const
 {
-	SceneElementProcessor::hashProcessedBound( path, context, h );
+	inPlug()->boundPlug()->hash( h );
 	isoValuePlug()->hash( h );
 }
 
-Imath::Box3f LevelSetToMesh::computeProcessedBound( const ScenePath &path, const Gaffer::Context *context, const Imath::Box3f &inputBound ) const
+Imath::Box3f LevelSetToMesh::computeProcessedObjectBound( const ScenePath &path, const Gaffer::Context *context ) const
 {
+	/// \todo `in.bound` includes the child bounds. Ideally we
+	/// would have a separate `in.objectBound` with just the bounds
+	/// of the input object.
+	const Box3f inputBound = inPlug()->boundPlug()->getValue();
 	const V3f offset( isoValuePlug()->getValue() );
 	return Box3f( inputBound.min - offset, inputBound.max + offset );
 }